% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragment-class.R
\docType{data}
\name{Fragment}
\alias{Fragment}
\title{ghql fragment class}
\value{
a \code{Fragment} class (R6 class)
}
\description{
ghql fragment class
}
\section{methods}{

\strong{Methods}
  \describe{
    \item{\code{fragment(name, x)}}{
     create a fragment by name
    }
  }
}
\examples{
\dontrun{
# make a fragment class
frag <- Fragment$new()

# define a fragment
frag$fragment('Watchers', '
  fragment on Repository {
    watchers(first: 3) {
      edges {
        node {
          name
       }
    }
  }
}')

# define another fragment
frag$fragment('Stargazers', '
  fragment on Repository {
    stargazers(first: 3) {
      edges {
        node {
          name
       }
    }
  }
}')
frag
frag$fragments
frag$fragments$Watchers
frag$fragments$Stargazers
}
}
\keyword{datasets}

