% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{graphql}
\alias{graphql}
\title{ghql client}
\usage{
graphql(url, headers)
}
\arguments{
\item{url}{(character) URL for the GraphQL schema}

\item{headers}{Any acceptable \pkg{httr} header, constructed typically
via \code{\link[httr]{add_headers}}. See examples}
}
\value{
a \code{GraphqlClient} class (R6 class)
}
\description{
ghql client
}
\section{methods}{

\strong{Methods}
  \describe{
    \item{\code{ping(...)}}{
     ping the GraphQL server, return HTTP status code
    }
    \item{\code{load_schema(schema_url, schema_file)}}{
     load schema, from URL or local file
    }
    \item{\code{dump_schema(file)}}{
     dump schema to a local file
    }
    \item{\code{schema2json(...)}}{
     convert schema to JSON
    }
    \item{\code{query(x)}}{
     define query in a character string
    }
    \item{\code{parse2json()}}{
     parse query string with libgraphqlparser and get back JSON
    }
  }
}
\examples{
\dontrun{
# make a client
# cli <- graphql(url = "https://api.github.com/graphql")

library("httr")
token <- Sys.getenv("GITHUB_GRAPHQL_TOKEN")
cli <- graphql(
  url = "https://api.github.com/graphql",
  headers = add_headers(Authorization = paste0("Bearer ", token))
)

# if the GraphQL server has a schema, you can load it
cli$load_schema()

# dump schema to local file
f <- tempfile(fileext = ".json")
cli$dump_schema(file = f)
readLines(f)
jsonlite::fromJSON(readLines(f))

# after dumping to file, you can later read schema from file for faster loading
rm(cli)
cli <- graphql(
  url = "https://api.github.com/graphql",
  headers = add_headers(Authorization = paste0("Bearer ", token))
)
cli$load_schema(schema_file = f)

# variables
cli$url
cli$schema
cli$schema$data
cli$schema$data$`__schema`
cli$schema$data$`__schema`$queryType
cli$schema$data$`__schema`$mutationType
cli$schema$data$`__schema`$subscriptionType
head(cli$schema$data$`__schema`$types)
cli$schema$data$`__schema`$directives

# methods
## ping - hopefully you get a 200
cli$ping()

## dump schema
cli$schema2json()

## define query
cli$query('query { }')
cli$query_string
### execute the query
cli$exec()

## another
cli$query('{
  viewer {
    repositories(last: 10, isFork: false, privacy: PUBLIC) {
      edges {
        node {
          isPrivate
          id
          name
        }
      }
    }
  }
}')
cli$query_string
### execute the query
cli$exec()
### parse query string to JSON (with libgraphqlparser)
(json <- cli$parse_2json())
jsonlite::fromJSON(json)
}
}

