% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{ql}
\alias{ql}
\title{Query Github using GraphQL}
\usage{
ql(query, ...)
}
\arguments{
\item{query}{(character) a GraphQL query}

\item{...}{Curl options passed to \code{\link[httr]{GET}}}
}
\value{
a list or tibble if possible
}
\description{
Query Github using GraphQL
}
\examples{
\dontrun{
jsonlite::fromJSON(ql('{"query": "query { }"}', config=verbose()))

x <- ql('{"query": "query { viewer { login }}"}', config=verbose())
jsonlite::fromJSON(x)

x <- ql('{"query": "query { viewer { login }}"}', config=verbose())
jsonlite::fromJSON(x)

qry <- '{"query": "query { search(type: ISSUE, query: "Octocat", first: 30) { edges { node { ... on Issue { title }}}}}"}'
ql(qry, config=verbose())

qry <- '{"query": "query { search(type: ISSUE, query: "Octocat", first: 30) { edges { node }}}"}'
ql(qry, config=verbose())

ql('{"query": "query {
  search(type: ISSUE, query: "Octocat", first: 30) {
    edges {
      node {
        ... on Issue {
          title
        }
      }
    }
  }
}"}')
}
}
\references{
\url{https://developer.github.com/early-access/graphql/}
}

