% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsl.R
\name{viewer}
\alias{viewer}
\title{Github GraphQL DSL}
\usage{
viewer(.args, ...)
}
\arguments{
\item{...}{Curl options passed to \code{\link[httr]{GET}}
\itemize{
 \item avatarURL (String) - A URL pointing to the user's public avatar.
 \item bio (String) - The user's public profile bio.
 \item company (String) - The user's public profile company.
 \item createdAt (DateTime) - When this user signed up.
 \item databaseId (Int) - Identifies the primary key from the database.
 \item email (String) - The user's public profile email.
 \item isBountyHunter (Boolean) - Whether or not this user is a participant in the GitHub Security Bug Bounty.
 \item isDeveloperProgramMember (Boolean) - Whether or not this user is a GitHub Developer Program member.
 \item isEmployee (Boolean) - Whether or not this user is a GitHub employee.
 \item isSiteAdmin (Boolean) - Whether or not this user is a site administrator.
 \item isViewer (Boolean) - Whether or not this user is the viewing user.
 \item location (String) - The user's public profile location.
 \item login (String) - The username used to login.
 \item name (String) - The user's public profile name.
 \item repository (Repository) - Find Repository.
 \item viewerCanFollow (Boolean) - Whether or not the viewer is able to follow the user.
 \item viewerIsFollowing (Boolean) - Whether or not this user is followed by the viewer.
 \item websiteURL (String) - A URL pointing to the user's public website/blog.
}}

\item{x}{(character) xxx}
}
\description{
Github GraphQL DSL
}
\examples{
viewer('.', login)
viewer('.', login, email)
viewer('.', login, email, watching(first = 10, name, isPrivate))
}
\keyword{internal}

