% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghql-package.R
\docType{package}
\name{ghql-package}
\alias{ghql-package}
\alias{ghql}
\title{ghql}
\description{
General purpose GraphQL client
}
\section{ghql API}{

The main interface in this package is [GraphqlClient], which produces
a client (R6 class) with various methods for interacting with a
GraphQL server. [GraphqlClient] also accepts various input parameters
to set a base URL, and any headers required, which is usually the required
set of things needed to connect to a GraphQL service.

[Query] is an interface to creating GraphQL queries,
which works together with [GraphqlClient]

[Fragment] is an interface to creating GraphQL fragments,
which works together with [GraphqlClient]
}

\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
